requirements.txt
chromadb
sentence-transformers
transformers
torch
flask
pyspellchecker
sentencepiece
